backend/requirements.txt
# Use more flexible version ranges to avoid conflicts
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-docx>=0.8.11
pandas>=2.1.3
pydantic>=2.5.0
requests>=2.31.0
python-multipart>=0.0.6
pytest>=7.4.3
sqlalchemy>=2.0.23
aiofiles>=23.2.1
requests==2.31.0
pandas==2.1.3
python-docx==1.1.0